feat: Improve delete logic
